refactor(team): extract SSE handlers + web UI proxy from broker.go
Phase 3 first slice. Two cohesive extractions out of broker.go:

(1) broker_web_proxy.go (199 LOC, new): the web UI server.
    - ServeWebUI: stands up the static web UI on its own port.
    - cacheControlMiddleware: long-cache hashed bundles, no-cache index.
    - webUIProxyHandler: same-origin proxy to the broker API with
      server-side Bearer token attachment.

(2) broker_sse.go (402 LOC, new): SSE streaming + tool-event audit.
    - handleEvents: per-broker event fanout (messages, actions, activity,
      office changes, wiki, notebook, entity, playbook, pam).
    - handleAgentStream: per-agent stdout SSE.
    - handleAgentToolEvent + recordAgentToolEvent: tool-call audit
      stream (drives the Hermes counter for skill_review nudges).
    - SkillCounter helpers: ensureSkillCounter, SetSkillCounter,
      skillCounterSummaryFromArgs.
    - formatAgentToolEvent + decodeToolEventField: SSE wire encoding.

  - internal/team/broker.go: -545 LOC (3404 -> 2859)
  - internal/team/broker_web_proxy.go: +199 LOC (new)
  - internal/team/broker_sse.go: +402 LOC (new)

No behavior change. internal/team tests pass with -race.

Cumulative broker.go reduction so far in Phase 3:
  3404 -> 2859 (web proxy + sse)   -545

Next slices: rate-limit, auth, wiki worker lifecycle, telemetry,
publish, defaults/state-normalization. Will land in subsequent stack PRs.
