+++ internal/team/broker.go
@@ -7,9 +7,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"log"
-	"log/slog"
 	"net"
 	"net/http"
 	"os"
@@ -21,7 +19,6 @@ import (
 	"sync/atomic"
 	"time"
 
-	wuphf "github.com/nex-crm/wuphf"
 	"github.com/nex-crm/wuphf/internal/brokeraddr"
 	"github.com/nex-crm/wuphf/internal/channel"
 	"github.com/nex-crm/wuphf/internal/company"
@@ -1106,556 +1103,14 @@ func (b *Broker) handleWebToken(w http.ResponseWriter, r *http.Request) {
 	_ = json.NewEncoder(w).Encode(map[string]string{"token": b.token})
 }
 
-func (b *Broker) handleEvents(w http.ResponseWriter, r *http.Request) {
-	if !b.requestHasBrokerAuth(r) {
-		http.Error(w, `{"error":"unauthorized"}`, http.StatusUnauthorized)
-		return
-	}
-
-	flusher, ok := w.(http.Flusher)
-	if !ok {
-		http.Error(w, "streaming unsupported", http.StatusInternalServerError)
-		return
-	}
-
-	w.Header().Set("Content-Type", "text/event-stream")
-	w.Header().Set("Cache-Control", "no-cache")
-	w.Header().Set("Connection", "keep-alive")
-	w.Header().Set("X-Accel-Buffering", "no")
-
-	messages, unsubscribeMessages := b.SubscribeMessages(256)
-	defer unsubscribeMessages()
-	actions, unsubscribeActions := b.SubscribeActions(256)
-	defer unsubscribeActions()
-	activity, unsubscribeActivity := b.SubscribeActivity(256)
-	defer unsubscribeActivity()
-	officeChanges, unsubscribeOffice := b.SubscribeOfficeChanges(64)
-	defer unsubscribeOffice()
-	wikiEvents, unsubscribeWiki := b.SubscribeWikiEvents(64)
-	defer unsubscribeWiki()
-	notebookEvents, unsubscribeNotebook := b.SubscribeNotebookEvents(64)
-	defer unsubscribeNotebook()
-	entityEvents, unsubscribeEntity := b.SubscribeEntityBriefEvents(64)
-	defer unsubscribeEntity()
-	factEvents, unsubscribeFacts := b.SubscribeEntityFactEvents(64)
-	defer unsubscribeFacts()
-	sectionsEvents, unsubscribeSections := b.SubscribeWikiSectionsUpdated(16)
-	defer unsubscribeSections()
-	playbookEvents, unsubscribePlaybook := b.SubscribePlaybookExecutionEvents(64)
-	defer unsubscribePlaybook()
-	playbookSynthEvents, unsubscribePlaybookSynth := b.SubscribePlaybookSynthesizedEvents(64)
-	defer unsubscribePlaybookSynth()
-	pamStarted, pamDone, pamFailed, unsubscribePam := b.SubscribePamActionEvents(64)
-	defer unsubscribePam()
-
-	writeEvent := func(name string, payload any) error {
-		data, err := json.Marshal(payload)
-		if err != nil {
-			return err
-		}
-		if _, err := fmt.Fprintf(w, "event: %s\ndata: %s\n\n", name, data); err != nil {
-			return err
-		}
-		flusher.Flush()
-		return nil
-	}
-
-	if err := writeEvent("ready", map[string]string{"status": "ok"}); err != nil {
-		return
-	}
-
-	heartbeat := time.NewTicker(15 * time.Second)
-	defer heartbeat.Stop()
-
-	for {
-		select {
-		case <-r.Context().Done():
-			return
-		case msg, ok := <-messages:
-			if !ok || writeEvent("message", map[string]any{"message": msg}) != nil {
-				return
-			}
-		case action, ok := <-actions:
-			if !ok || writeEvent("action", map[string]any{"action": action}) != nil {
-				return
-			}
-		case snapshot, ok := <-activity:
-			if !ok || writeEvent("activity", map[string]any{"activity": snapshot}) != nil {
-				return
-			}
-		case evt, ok := <-officeChanges:
-			if !ok || writeEvent("office_changed", evt) != nil {
-				return
-			}
-		case evt, ok := <-wikiEvents:
-			if !ok || writeEvent("wiki:write", evt) != nil {
-				return
-			}
-		case evt, ok := <-notebookEvents:
-			if !ok || writeEvent("notebook:write", evt) != nil {
-				return
-			}
-		case evt, ok := <-entityEvents:
-			if !ok || writeEvent("entity:brief_synthesized", evt) != nil {
-				return
-			}
-		case evt, ok := <-factEvents:
-			if !ok || writeEvent("entity:fact_recorded", evt) != nil {
-				return
-			}
-		case evt, ok := <-sectionsEvents:
-			if !ok || writeEvent(wikiSectionsEventName, evt) != nil {
-				return
-			}
-		case evt, ok := <-playbookEvents:
-			if !ok || writeEvent("playbook:execution_recorded", evt) != nil {
-				return
-			}
-		case evt, ok := <-playbookSynthEvents:
-			if !ok || writeEvent("playbook:synthesized", evt) != nil {
-				return
-			}
-		case evt, ok := <-pamStarted:
-			if !ok || writeEvent("pam:action_started", evt) != nil {
-				return
-			}
-		case evt, ok := <-pamDone:
-			if !ok || writeEvent("pam:action_done", evt) != nil {
-				return
-			}
-		case evt, ok := <-pamFailed:
-			if !ok || writeEvent("pam:action_failed", evt) != nil {
-				return
-			}
-		case <-heartbeat.C:
-			if _, err := fmt.Fprintf(w, ": ping\n\n"); err != nil {
-				return
-			}
-			flusher.Flush()
-		}
-	}
-}
-
-// handleAgentToolEvent appends a tool-call log line to the agent's stream so
-// the per-agent activity panel shows which MCP tool was invoked with what
-// arguments. Without this, the stream only shows raw pane-captured stdout —
-// useless for agents whose work happens entirely through MCP tool calls.
-//
-// Body: {"slug":"ceo","phase":"call|result|error","tool":"team_broadcast","args":"...","result":"...","error":"..."}
-// Phase is informational; all fields but slug are optional.
-func (b *Broker) handleAgentToolEvent(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodPost {
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-	var body struct {
-		Slug   string `json:"slug"`
-		Phase  string `json:"phase,omitempty"`
-		Tool   string `json:"tool,omitempty"`
-		Args   string `json:"args,omitempty"`
-		Result string `json:"result,omitempty"`
-		Error  string `json:"error,omitempty"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		http.Error(w, "invalid json", http.StatusBadRequest)
-		return
-	}
-	slug := strings.TrimSpace(body.Slug)
-	if slug == "" {
-		http.Error(w, "missing slug", http.StatusBadRequest)
-		return
-	}
-	stream := b.AgentStream(slug)
-	if stream != nil {
-		line := formatAgentToolEvent(body.Phase, body.Tool, body.Args, body.Result, body.Error)
-		if line != "" {
-			stream.Push(line)
-		}
-	}
-
-	// Drive the Hermes counter. We only count once per tool call so we
-	// branch on phase=="call" — the call/result/error fan-out from a
-	// single tool invocation must NOT triple-count. team_skill_create /
-	// team_skill_patch reset the tally; everything else increments and
-	// may fire a skill_review_nudge task.
-	b.recordAgentToolEvent(slug, body.Phase, body.Tool, body.Args)
-
-	w.WriteHeader(http.StatusOK)
-}
-
-// recordAgentToolEvent updates the broker's SkillCounter for one
-// tool-event payload. It is split out from handleAgentToolEvent so tests
-// can drive the counter directly without going through HTTP, and so the
-// b.mu acquisition for the nudge task creation is centralized in one
-// place.
-//
-// We only act on phase=="call" — the call/result/error fan-out per tool
-// invocation must not triple-count. Empty phase is treated as "call" for
-// backward compatibility; older agents that didn't tag the phase still
-// only post one event per call.
-func (b *Broker) recordAgentToolEvent(slug, phase, tool, args string) {
-	slug = strings.TrimSpace(slug)
-	tool = strings.TrimSpace(tool)
-	if slug == "" || tool == "" {
-		return
-	}
-	phase = strings.TrimSpace(phase)
-	if phase != "" && phase != "call" {
-		// result / error events from the same call — already counted.
-		return
-	}
-	counter := b.ensureSkillCounter()
-	if counter == nil {
-		return
-	}
-
-	// Skill-authoring tools reset the counter (the agent just codified
-	// something). Everything else increments.
-	if IsSkillAuthoringTool(tool) {
-		counter.Reset(slug)
-		return
-	}
-
-	summary := skillCounterSummaryFromArgs(tool, args)
-	shouldNudge, _ := counter.Increment(slug, tool, summary)
-	if !shouldNudge {
-		return
-	}
-
-	b.mu.Lock()
-	taskID, err := b.fireSkillReviewNudgeLocked(slug)
-	if err == nil {
-		atomic.AddInt64(&b.skillCompileMetrics.CounterNudgesFiredTotal, 1)
-		if saveErr := b.saveLocked(); saveErr != nil {
-			slog.Warn("skill_counter_nudge_persist_failed",
-				"agent", slug, "task_id", taskID, "err", saveErr)
-		}
-	}
-	b.mu.Unlock()
-	if err != nil {
-		slog.Warn("skill_counter_nudge_create_failed",
-			"agent", slug, "err", err)
-	}
-}
-
-// ensureSkillCounter lazily constructs the SkillCounter. Like the other
-// skill-* singletons, the counter is built on first use so tests that
-// never trigger an agent tool call pay no cost.
-func (b *Broker) ensureSkillCounter() *SkillCounter {
-	b.mu.Lock()
-	if b.skillCounter != nil {
-		c := b.skillCounter
-		b.mu.Unlock()
-		return c
-	}
-	c := NewSkillCounter()
-	b.skillCounter = c
-	b.mu.Unlock()
-	return c
-}
-
-// SetSkillCounter replaces the broker's counter — used by tests to inject
-// a counter with a specific threshold / cooldown / clock without going
-// through env vars.
-func (b *Broker) SetSkillCounter(c *SkillCounter) {
-	b.mu.Lock()
-	b.skillCounter = c
-	b.mu.Unlock()
-}
-
-// skillCounterSummaryFromArgs renders a one-line summary of a tool call
-// for the per-agent ring buffer. The args field is a JSON-encoded
-// string (as posted by the MCP client), so we don't try to parse it —
-// we just trim and truncate. Real human review of the nudge task uses
-// the agent's own activity stream for full detail.
-func skillCounterSummaryFromArgs(tool, args string) string {
-	args = strings.TrimSpace(args)
-	if args == "" {
-		return tool
-	}
-	// Collapse whitespace runs so the summary stays single-line.
-	args = strings.Join(strings.Fields(args), " ")
-	return truncateSummary(args, 120)
-}
-
-// formatAgentToolEvent renders one structured audit record for the per-agent
-// stream. SSE data lines must stay single-line; JSON encoding preserves exact
-// arguments/results while escaping embedded newlines.
-func formatAgentToolEvent(phase, tool, args, result, errStr string) string {
-	tool = strings.TrimSpace(tool)
-	phase = strings.TrimSpace(phase)
-	if phase == "" {
-		phase = "tool"
-	}
-	if tool == "" {
-		return ""
-	}
-	payload := map[string]any{
-		"type":  "mcp_tool_event",
-		"phase": phase,
-		"tool":  tool,
-	}
-	if args != "" {
-		payload["arguments"] = decodeToolEventField(args)
-	}
-	if result != "" {
-		payload["result"] = decodeToolEventField(result)
-	}
-	if errStr != "" {
-		payload["error"] = decodeToolEventField(errStr)
-	}
-	data, err := json.Marshal(payload)
-	if err != nil {
-		return ""
-	}
-	return string(data)
-}
-
-func decodeToolEventField(raw string) any {
-	raw = strings.TrimSpace(raw)
-	if raw == "" {
-		return ""
-	}
-	var decoded any
-	if err := json.Unmarshal([]byte(raw), &decoded); err == nil {
-		return decoded
-	}
-	return raw
-}
-
-// handleAgentStream serves a per-agent stdout SSE stream.
-// Recent lines are replayed as initial history, then new lines are pushed live.
-// Path: /agent-stream/{slug}
-func (b *Broker) handleAgentStream(w http.ResponseWriter, r *http.Request) {
-	slug := strings.TrimPrefix(r.URL.Path, "/agent-stream/")
-	if slug == "" {
-		http.Error(w, "missing agent slug", http.StatusBadRequest)
-		return
-	}
-	flusher, ok := w.(http.Flusher)
-	if !ok {
-		http.Error(w, "streaming unsupported", http.StatusInternalServerError)
-		return
-	}
-	w.Header().Set("Content-Type", "text/event-stream")
-	w.Header().Set("Cache-Control", "no-cache")
-	w.Header().Set("Connection", "keep-alive")
-	w.Header().Set("X-Accel-Buffering", "no")
-
-	stream := b.AgentStream(slug)
-
-	// Replay recent history so the client sees context immediately.
-	history := stream.recent()
-	for _, line := range history {
-		if _, err := fmt.Fprintf(w, "data: %s\n\n", line); err != nil {
-			return
-		}
-	}
-	// If no history, send a connected event so the client knows the stream is live.
-	if len(history) == 0 {
-		if _, err := fmt.Fprintf(w, "data: [connected]\n\n"); err != nil {
-			return
-		}
-	}
-	flusher.Flush()
-
-	lines, unsubscribe := stream.subscribe()
-	defer unsubscribe()
-
-	heartbeat := time.NewTicker(15 * time.Second)
-	defer heartbeat.Stop()
-
-	for {
-		select {
-		case <-r.Context().Done():
-			return
-		case line, ok := <-lines:
-			if !ok {
-				return
-			}
-			if _, err := fmt.Fprintf(w, "data: %s\n\n", line); err != nil {
-				return
-			}
-			flusher.Flush()
-		case <-heartbeat.C:
-			if _, err := fmt.Fprintf(w, ": ping\n\n"); err != nil {
-				return
-			}
-			flusher.Flush()
-		}
-	}
-}
+// SSE handlers and the tool-event audit channel (handleEvents,
+// handleAgentStream, handleAgentToolEvent, recordAgentToolEvent,
+// SkillCounter helpers) moved to broker_sse.go.
 
 // ServeWebUI starts a static file server for the web UI on the given port.
 // Returns an error if the port cannot be bound (e.g. already in use).
-func (b *Broker) ServeWebUI(port int) error {
-	b.webUIOrigins = []string{
-		fmt.Sprintf("http://localhost:%d", port),
-		fmt.Sprintf("http://127.0.0.1:%d", port),
-	}
-
-	// Resolution order for the web UI assets:
-	//   1. filesystem web/dist/ (local dev after `npm run build`)
-	//   2. embedded FS (single-binary installs via curl | bash)
-	exePath, _ := os.Executable()
-	webDir := filepath.Join(filepath.Dir(exePath), "web")
-	if _, err := os.Stat(webDir); os.IsNotExist(err) {
-		webDir = "web"
-	}
-	var fileServer http.Handler
-	distDir := filepath.Join(webDir, "dist")
-	distIndex := filepath.Join(distDir, "index.html")
-	if _, err := os.Stat(distIndex); err == nil {
-		// Real Vite build output on disk — use it.
-		fileServer = http.FileServer(http.Dir(distDir))
-	} else if embeddedFS, ok := wuphf.WebFS(); ok {
-		// No on-disk build; use embedded assets.
-		fileServer = http.FileServer(http.FS(embeddedFS))
-	} else {
-		// Nothing available; serve webDir as-is so 404s come from the actual FS.
-		fileServer = http.FileServer(http.Dir(webDir))
-	}
-	mux := http.NewServeMux()
-	brokerURL := brokeraddr.ResolveBaseURL()
-	if addr := strings.TrimSpace(b.Addr()); addr != "" {
-		brokerURL = "http://" + addr
-	}
-	// Same-origin proxy to the broker for app API routes and onboarding wizard routes.
-	// Both are wrapped in webUIRebindGuard: the proxy auto-attaches the broker's
-	// Bearer token server-side, so without a Host/RemoteAddr check, a DNS-rebinding
-	// attack against an attacker-controlled hostname that resolves to 127.0.0.1
-	// would ride the token and control the entire office.
-	mux.Handle("/api/", webUIRebindGuard(b.webUIProxyHandler(brokerURL, "/api")))
-	mux.Handle("/onboarding/", webUIRebindGuard(b.webUIProxyHandler(brokerURL, "")))
-	// Token endpoint — no auth needed, but we require a same-origin loopback request.
-	// Otherwise this endpoint leaks the broker bearer to any browser page that
-	// can reach the web UI port via DNS rebinding.
-	mux.Handle("/api-token", webUIRebindGuard(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		_ = json.NewEncoder(w).Encode(map[string]string{
-			"token":      b.token,
-			"broker_url": brokerURL,
-		})
-	})))
-	// Cache policy: index.html must be re-fetched every load so users pick up
-	// new JS/CSS bundle hashes immediately after an upgrade. Hashed assets
-	// under /assets/ are content-addressed and safe to cache aggressively.
-	// Without this, users stay pinned to a stale bundle for days because
-	// Chrome's heuristic cache revalidates HTML only occasionally.
-	// Serve generated images from ~/.wuphf/office/artist/ so the BoardRoom
-	// can render them inline via standard markdown <img>. Browsers can't
-	// fetch file:// URLs and don't carry the broker's bearer token on
-	// <img> requests, so this mount must live on the web-UI port (no auth)
-	// rather than the API mux. Path traversal is bounded by http.FileServer
-	// + http.Dir; we strip the prefix so requests resolve relative to the
-	// artist root.
-	artistRoot := imagegenArtistRoot()
-	mux.Handle("/artist-files/", http.StripPrefix(
-		"/artist-files/",
-		http.FileServer(http.Dir(artistRoot)),
-	))
-
-	mux.Handle("/", cacheControlMiddleware(fileServer))
-	addr := fmt.Sprintf("127.0.0.1:%d", port)
-	// noctx: net.Listen is the blocking primitive; the lint rule is meant
-	// for HTTP clients. Use ListenConfig.Listen with a Background context
-	// so the linter's intent (no caller-controllable cancellation lost) is
-	// satisfied without changing the actual lifecycle.
-	lc := &net.ListenConfig{}
-	ln, err := lc.Listen(context.Background(), "tcp", addr)
-	if err != nil {
-		return fmt.Errorf("web UI: listen on %s: %w", addr, err)
-	}
-	srv := &http.Server{Handler: mux}
-	go func() {
-		if err := srv.Serve(ln); err != nil && !errors.Is(err, http.ErrServerClosed) {
-			log.Printf("broker web UI proxy: serve on :%d: %v", port, err)
-		}
-	}()
-	return nil
-}
-
-// cacheControlMiddleware sets conservative cache headers on the web UI so
-// clients always receive fresh HTML and mutable assets, while long-cached
-// hashed bundles under /assets/ stay immutable for efficiency.
-func cacheControlMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		path := r.URL.Path
-		switch {
-		case strings.HasPrefix(path, "/assets/"):
-			// Vite bundles hashed filenames; they never change for a given URL.
-			w.Header().Set("Cache-Control", "public, max-age=31536000, immutable")
-		default:
-			// Everything else (index.html, themes/*.css, favicons that share a
-			// stable path) must re-validate on each load so upgrades land
-			// immediately.
-			w.Header().Set("Cache-Control", "no-cache, must-revalidate")
-		}
-		next.ServeHTTP(w, r)
-	})
-}
-
-func (b *Broker) webUIProxyHandler(brokerURL, stripPrefix string) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		targetPath := r.URL.Path
-		if stripPrefix != "" {
-			targetPath = strings.TrimPrefix(targetPath, stripPrefix)
-		}
-		if targetPath == "" {
-			targetPath = "/"
-		}
-		target := brokerURL + targetPath
-		if r.URL.RawQuery != "" {
-			target += "?" + r.URL.RawQuery
-		}
-
-		proxyReq, err := http.NewRequestWithContext(r.Context(), r.Method, target, r.Body)
-		if err != nil {
-			http.Error(w, "proxy error", http.StatusBadGateway)
-			return
-		}
-		setProxyClientIPHeaders(proxyReq.Header, r.RemoteAddr)
-		proxyReq.Header.Set("Authorization", "Bearer "+b.token)
-		proxyReq.Header.Set("Content-Type", r.Header.Get("Content-Type"))
-
-		client := http.DefaultClient
-		if r.Header.Get("Accept") == "text/event-stream" {
-			client = &http.Client{Timeout: 0}
-		}
-		resp, err := client.Do(proxyReq)
-		if err != nil {
-			http.Error(w, "broker unreachable", http.StatusBadGateway)
-			return
-		}
-		defer resp.Body.Close()
-
-		for k, v := range resp.Header {
-			for _, vv := range v {
-				w.Header().Add(k, vv)
-			}
-		}
-		w.WriteHeader(resp.StatusCode)
-
-		if resp.Header.Get("Content-Type") == "text/event-stream" {
-			flusher, canFlush := w.(http.Flusher)
-			buf := make([]byte, 4096)
-			for {
-				n, readErr := resp.Body.Read(buf)
-				if n > 0 {
-					w.Write(buf[:n]) //nolint:errcheck
-					if canFlush {
-						flusher.Flush()
-					}
-				}
-				if readErr != nil {
-					break
-				}
-			}
-			return
-		}
-		_, _ = io.Copy(w, resp.Body)
-	})
-}
+// Web UI server (ServeWebUI, cacheControlMiddleware, webUIProxyHandler)
+// moved to broker_web_proxy.go.
 
 // Messages returns all channel messages (for the Go TUI channel view).
 func (b *Broker) Messages() []channelMessage {
